requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
feedparser==6.0.10
sqlalchemy==2.0.23
aiosqlite==0.19.0
apscheduler==3.10.4
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
urllib3==2.1.0